Refaktor. Skal lage en egen loader funksjon som kan ta inn ytelse. Mininerer duplikat kode
